requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.0
pydantic==2.6.0
pytest==8.1.0
httpx==0.27.0
python-dotenv-1.2.2